player-stats-and-power-ups-update

+++ Assets/Resources/Scripts/Enemy/EnemyHealthManager.cs
@@ -5,19 +5,19 @@ namespace Enemy
     public class EnemyHealthManager : MonoBehaviour
     {
         public int MaxHealth = 200;
-        public int CurrentHealth;
+        public float CurrentHealth;
         
         private void Start()
         {
             CurrentHealth = MaxHealth;
         }
         
-        public void TakeDamage(int damage)
+        public void TakeDamage(float damage)
         {
             CurrentHealth -= damage;
         }
 
-        public int GetCurrentHp()
+        public float GetCurrentHp()
         {
             return CurrentHealth;
         }
